Exit with the return code from command.parse
Command.parse returns a non-zero return code in a variety of
circumstances. However, the func command line script (script/func) was
ignoring that return code, and thus was hiding a number of error
conditions.

This patch uses sys.exit to return the return codes from command.parse
to the command line user.